backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-decouple==3.8
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
boto3==1.34.0
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
minio==7.2.0
PyJWT==2.8.0
